Client Registration: Transfers command handler dispatch function to appropriate file #3
